requirements.txt
Flask==3.0.0
boto3==1.34.6
pdfplumber==0.11.5
python-docx==1.1.2
gspread==6.2.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
requests==2.31.0
google-api-core==2.24.2 
google-api-python-client==2.163.0 
proto-plus==1.26.1
uritemplate==4.1.1